pyproject.toml
[project]
name = "whatsapp-wrapped"
version = "0.1.0"
description = "WhatsApp Wrapped - Chat analytics and stats"
readme = "README.md"
requires-python = ">=3.11"

dependencies = [
    # Shared (CLI + Backend)
    "pandas>=2.0",
    "numpy>=1.24",
    "emoji>=2.0",

    # CLI only
    "rich>=14.0",
    "plotly>=6.0",

    # Backend
    "flask>=3.0",
    "gunicorn>=21.0",
    "celery[redis]>=5.3",
    "redis>=5.0",
    "sqlalchemy>=2.0",
    "psycopg2-binary>=2.9",
    "flask-sqlalchemy>=3.1",
    "boto3>=1.34",
    "flask-cors>=4.0",
    "flask-limiter[redis]>=3.0",
    "python-magic>=0.4",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-flask>=1.3",
    "ruff>=0.1",
    "httpx>=0.27",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
